feat: display queue from backend and make queue more modular
