+++ src/Router/Router.jsx
@@ -2,7 +2,7 @@ import About from '../About/About';
 import Container from '@material-ui/core/Container'
 import Home from '../Home/Home';
 import PatientTable from '../Patients/PatientTable';
-import Queue from '../Queue/Queue';
+import QueueContainer from '../Queue/QueueContainer';
 
 import { Routes, Route } from 'react-router-dom';
 
@@ -11,7 +11,7 @@ const MyRouter = () => {
     <Routes>
       <Route path="/" element={ <Home /> } />
       <Route path="about" element={ <About /> } />
-      <Route path="queue" element={ <Queue /> } />
+      <Route path="queue" element={ <QueueContainer /> } />
       <Route path="patients" 
         element={ 
           <Container
